Version Update for java 17 and spring-boot 3.2 (#28)

+++ src/main/java/org/sunbird/workflow/service/impl/WorkflowServiceImpl.java
@@ -929,7 +929,7 @@ public class WorkflowServiceImpl implements Workflowservice {
 			for (Object[] result : resultSet) {
 				WfStatusCountDTO dto = new WfStatusCountDTO();
 				dto.setCurrentStatus((String) result[0]);
-				dto.setStatusCount(((BigInteger) result[1]).longValue());
+				dto.setStatusCount(((Long) result[1]));
 				statusCountDTOs.add(dto);
 			}
 			localCache.put(applicationId,statusCountDTOs);
